Add files via upload
Random documents scoured on the internet regarding the origin, calculation, and validity of MMLs and W. D. Gann's core principles. The one I found most interesting is "MurreyMathCalculator, Version 1.3 (Dec 1999).xls" which gives a rudimentary formula of how the Murrey Math Lines are supposed to be calculated. It is in line with more sources than other formulas.

Source for most of these docs: http://www.tradewithme.me/trading-systems-iv--murray-maths.html