Update LICENSE
Extremely permissive (no restrictions at all) when redist. as open source. Must attribute if closed source.
This license should supersede MPL2 as it was written prior. It's also implied that this license can be changed, so really this is a license that says "I would appreciate it if you would put me in your acknowledgements section".